Restructure ModelIterator code to conform better with other tools
Moved the code to /src and the example to /example

Also did some other small modifications the example:
* Forward comparison of the ModelIterator pseudo_val to the value
  to enable comparing to anything the value can be compared to.
* Get rid of QStringLiteral where we used it compare against
* Added a few comments

Reviewed-on: https://codereview.kdab.com/70814
Tested-by: Continuous Integration <build@kdab.com>
